repo: phase 3 - remove all references to deprecated code
