+++ facedancer/backends/moondancer.py
@@ -10,8 +10,6 @@ import traceback
 
 from ..core import *
 from ..constants import DeviceSpeed
-from ..deprecated.USB import *
-from ..deprecated.USBEndpoint import USBEndpoint
 
 # add a TRACE level to logging
 logging.TRACE = 5
